+++ js/name.js
@@ -1,0 +1,14 @@
+const NAME = ['Антон',
+  'Иван',
+  'Андрей',
+  'Алина',
+  'Юлия',
+  'Екатерина',
+  'Денис',
+  'Владислав',
+  'Ольга',
+  'Алиса',
+  'Анастасия'
+];
+
+export {NAME};
