+++ js/util.js
@@ -1,0 +1,13 @@
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = function (elements) {
+  return elements[getRandomInteger(0,elements.length - 1)];
+};
+
+export{getRandomArrayElement};
+export{getRandomInteger};
